Update enable-support-tls-environment.md
The Select-MgProfile and Get-MgProfile cmdlets were used in earlier versions of the Microsoft Graph PowerShell module to switch between and retrieve the current profile, respectively. However, with the release of version 2.0 and above, these cmdlets are no longer relevant or needed.

In the newer versions of the Microsoft Graph PowerShell module (v2.0 and above), the cmdlets for the beta endpoint have been separated into their own standalone modules, prefixed with Microsoft.Graph.Beta. This separation eliminates the need for switching profiles within the same module namespace.